requirements.txt
Django==4.0.1
django-autoslug==1.9.8
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
Pillow==9.0.0
whitenoise==5.3.0
